requirements.txt
flask==3.0.0
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.3
werkzeug==3.0.1
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.0
Jinja2==3.1.2
